refactoring provenance class definitions into their own module, checking file existance before calls to add provenance, removed set() copy constructor
